Remove old cluster code
