--- cluster/cluster.go
@@ -1,1 +1,0 @@
-package cluster // import "github.com/influxdata/influxdb/cluster"
