--- cluster/internal/data.proto
@@ -1,104 +1,0 @@
-package cluster;
-
-message WriteShardRequest {
-    required uint64 ShardID = 1;
-    repeated bytes  Points  = 2;
-    optional string Database = 3;
-    optional string RetentionPolicy = 4;
-}
-
-message WriteShardResponse {
-    required int32  Code    = 1;
-    optional string Message = 2;
-}
-
-message ExecuteStatementRequest {
-    required string Statement = 1;
-    required string Database  = 2;
-}
-
-message ExecuteStatementResponse {
-    required int32  Code    = 1;
-    optional string Message = 2;
-}
-
-message CreateIteratorRequest {
-    repeated uint64 ShardIDs = 1;
-    required bytes  Opt      = 2;
-}
-
-message CreateIteratorResponse {
-    optional string Err  = 1;
-    required int32  Type = 2;
-
-    optional IteratorStats Stats = 3;
-}
-
-message IteratorStats {
-    optional int64 SeriesN = 1;
-    optional int64 PointN  = 2;
-}
-
-message FieldDimensionsRequest {
-    repeated uint64 ShardIDs = 1;
-    required bytes  Sources  = 2;
-}
-
-message FieldDimensionsResponse {
-    repeated string Fields     = 1;
-    repeated string Dimensions = 2;
-    optional string Err        = 3;
-}
-
-message SeriesKeysRequest {
-    repeated uint64 ShardIDs = 1;
-    required bytes  Opt      = 2;
-}
-
-message SeriesKeysResponse {
-    optional bytes  SeriesList = 1;
-    optional string Err        = 2;
-}
-
-message ExpandSourcesRequest {
-    repeated uint64 ShardIDs = 1;
-    required bytes  Sources  = 2;
-}
-
-message ExpandSourcesResponse {
-    required bytes  Sources  = 1;
-    optional string Err      = 2;
-}
-
-message RemoteMonitorRequest {
-    repeated string RemoteAddrs = 1;
-    required string NodeID      = 2;
-    required string Username    = 3;
-    required string Password    = 4;
-    required uint64 ClusterID   = 5;
-}
-
-message RemoteMonitorResponse {
-    optional string Err = 1;
-}
-
-message BackupShardRequest {
-    required uint64 ShardID = 1;
-    optional int64  Since   = 2;
-}
-
-message BackupShardResponse {
-    optional string Err = 2;
-}
-
-message CopyShardRequest {
-    required string Host     = 1;
-    required string Database = 2;
-    required string Policy   = 3;
-    required uint64 ShardID  = 4;
-    optional int64  Since    = 5;
-}
-
-message CopyShardResponse {
-    optional string Err = 2;
-}
